Refactored User Types to consolidate.
