+++ src/hooks/useAuth.ts
@@ -1,15 +1,6 @@
 import { useContext } from 'react';
 import { AuthContext } from '../contexts/AuthContext';
-
-export interface User {
-  id: string;
-  name: string;
-}
-
-export interface AuthContextType {
-  user: User | null;
-  isAuthenticated: boolean;
-}
+import type { AuthContextType } from '../types';
 
 export const useAuth = () => {
   const context = useContext(AuthContext);
@@ -17,4 +8,4 @@ export const useAuth = () => {
     throw new Error('useAuth must be used within an AuthProvider');
   }
   return context;
-}; 
+};
