+++ src/types/index.ts
@@ -1,0 +1,5 @@
+export * from './auth.types';
+export * from './user.types';
+export * from './note.types';
+export * from './quiz.types';
+export * from './course.types';
