+++ src/types/note.types.ts
@@ -1,0 +1,8 @@
+export interface Note {
+  id: string;
+  courseId: string;
+  unitName: string;
+  lessonName: string;
+  text: string;
+  updatedAt: string;
+}
